+++ packages/components/lib/layout/layout.dart
@@ -2,6 +2,8 @@ export 'ui_box.dart';
 export 'ui_form_field.dart';
 export 'ui_gap.dart';
 export 'ui_inset.dart';
+export 'ui_list.dart';
+export 'ui_page.dart';
 export 'ui_responsive.dart';
 export 'ui_spacer.dart';
 export 'ui_spacing.dart';